Implement Phases 4-5: rendering pipeline, input system, and panel tree
Add CPU software rasterizer (Painter) with canvas blending, coordinate
transforms, clip rects, and built-in 5x7 bitmap font. Add tile cache with
LRU eviction and structural wgpu compositor stub. Implement arena-based
panel tree with PanelBehavior trait, PanelCtx scoped mutation API, View
with visit stack navigation, view animators (kinetic/speeding/visiting),
and input filters. Add input system with key/mouse/touch state tracking,
hotkey parsing, and cursor types.

Fix canvas_blend overflow by widening arithmetic from i16 to i32.

+++ src/foundation/color.rs
@@ -113,9 +113,9 @@ impl Color {
     /// `self` is the current target pixel, `source` is the color being painted,
     /// `canvas` is the background canvas color, `alpha` is blend strength (0–255).
     pub fn canvas_blend(self, source: Color, canvas: Color, alpha: u8) -> Color {
-        let a = alpha as i16;
+        let a = alpha as i32;
         let blend_ch = |target: u8, src: u8, cvs: u8| -> u8 {
-            let result = target as i16 + ((src as i16 - cvs as i16) * a) / 255;
+            let result = target as i32 + ((src as i32 - cvs as i32) * a) / 255;
             result.clamp(0, 255) as u8
         };
         Color::rgba(
